chore: remove asm-logs crate
