--- crates/asm/logs/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "strata-asm-logs"
-version = "0.1.0"
-edition = "2021"
-
-[lints]
-workspace = true
-
-[dependencies]
-strata-asm-common.workspace = true
-strata-checkpoint-types.workspace = true
-strata-checkpoint-types-ssz.workspace = true
-strata-codec.workspace = true
-strata-codec-utils.workspace = true
-strata-identifiers.workspace = true
-strata-msg-fmt.workspace = true
-strata-predicate.workspace = true
-strata-primitives.workspace = true
-
-[dev-dependencies]
-proptest.workspace = true
-strata-checkpoint-types-ssz = { workspace = true, features = ["test-utils"] }
-strata-identifiers = { workspace = true, features = ["arbitrary"] }
-strata-test-utils.workspace = true
