--- crates/asm/logs/src/lib.rs
@@ -1,17 +1,0 @@
-//! # ASM Log Types
-//!
-//! This crate provides structured log types for the Anchor State Machine (ASM) in the Strata
-//! protocol. It defines various log entry types that capture important events within the system.
-
-pub mod asm_stf;
-pub mod checkpoint;
-pub mod constants;
-pub mod deposit;
-pub mod export;
-pub mod forced_inclusion;
-
-pub use asm_stf::AsmStfUpdate;
-pub use checkpoint::{CheckpointTipUpdate, CheckpointUpdate};
-pub use deposit::DepositLog;
-pub use export::NewExportEntry;
-pub use forced_inclusion::ForcedInclusionData;
