+++ crates/asm/subprotocols/checkpoint-v0/Cargo.toml
@@ -14,8 +14,10 @@ strata-asm-common.workspace = true
 strata-asm-logs.workspace = true
 strata-asm-txs-checkpoint-v0.workspace = true
 strata-checkpoint-types.workspace = true
-strata-checkpoint-types-ssz-asm = { package = "strata-checkpoint-types-ssz", git = "https://github.com/alpenlabs/asm", rev = "49c381ec0f3e11a32d3aa20a521ee49270662b21" }
+strata-codec.workspace = true
+strata-codec-utils.workspace = true
 strata-identifiers.workspace = true
+strata-msg-fmt.workspace = true
 strata-ol-bridge-types.workspace = true
 strata-params.workspace = true
 strata-predicate.workspace = true
